+++ websocket/client.go
@@ -11,18 +11,6 @@ import (
 	"github.com/thecoderstudio/apollo-agent/oauth"
 )
 
-// ShellIO is used for communicating shell input, output and error streams.
-type ShellIO struct {
-	ConnectionID string `json:"connection_id"`
-	Message      string `json:"message"`
-}
-
-// Command is used to instruct the agent to execute a pre-defined command.
-type Command struct {
-    ConnectionID    string `json:"connection_id"`
-    Command         string `json:"command"`
-}
-
 // Connection specifies the interface for client connection
 type Connection interface {
 	Close() error
